requirements.txt
awsebcli==3.14.8
botocore==1.12.71
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
docutils==0.14
httpie==1.0.2
idna==2.8
jmespath==0.9.3
pathspec==0.5.5
Pillow==5.3.0
Pygments==2.3.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
urllib3==1.22
boto3==1.9.71
botocore==1.12.71
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1.4
django-cors-headers==2.4.0
django-storages==1.7.1
djangorestframework==3.9.0
docutils==0.14
httpie==1.0.2
idna==2.8
jmespath==0.9.3
Pillow==5.3.0
Pygments==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
boto3==1.9.71
botocore==1.12.71
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1.4
django-cors-headers==2.4.0
django-storages==1.7.1
djangorestframework==3.9.0
docutils==0.14
httpie==1.0.2
idna==2.8
jmespath==0.9.3
Pillow==5.3.0
Pygments==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
